refactor(analyzer): extract docstring/signature helpers into call-graph-extract.ts (slice 3)
Next safe slice of modularize-call-graph-builder. Moves the DOCSTRING / SIGNATURE
EXTRACTION HELPERS section (extractDocstringBefore, extractDeclaration) out of
call-graph.ts into call-graph-extract.ts — two pure string-scanning functions with
zero dependency on the rest of the analyzer (no tree-sitter, no module state).

They were file-internal helpers (never on call-graph.ts's public surface), so they
are imported back, NOT re-exported — the public import surface is unchanged.
Taken before slice 2 as the safest small slice. call-graph.ts: 5,150 → 4,951 lines.

Behavior-preserving, verified four ways:
- Snapshot oracle FIRST strengthened to serialize each node's docstring + signature
  (fixture now has a JSDoc'd TS fn, an exported async fn, and a Python docstring +
  annotated signature, so both moved functions are exercised across TS and Python),
  then captured before/after: IDENTICAL — SHA-256 58107ac0…
- Public export surface byte-identical (multi-line-aware diff: 0 added, 0 removed).
- tsc compiles all importers clean; build/lint/typecheck green.
- Full suite green: 279 files / 5534 tests.

No feature, no dependency, no LLM, no persisted artifact.
